Move shared functionality for UI projects to the shared-ui library
