+++ apps/web-app/src/app/core/config/app.config.ts
@@ -1,8 +1,8 @@
 import { provideHttpClient } from '@angular/common/http';
 import { ApplicationConfig, provideBrowserGlobalErrorListeners, provideZonelessChangeDetection } from '@angular/core';
 import { provideRouter } from '@angular/router';
+import { provideClientConfig } from '@craftsmans-ledger/shared-ui';
 import { appRoutes } from './app.routes';
-import { provideClientConfig } from './config.service';
 
 export const appConfig: ApplicationConfig = {
     providers: [
