+++ apps/web-app/src/app/core/config/index.ts
@@ -1,2 +1,1 @@
 export * from './app.config';
-export * from './config.service';
