+++ apps/web-app/src/app/index.ts
@@ -1,3 +1,1 @@
 export * from './core';
-export * from './items';
-export * from './shared';
