+++ apps/web-app/src/app/shared/index.ts
@@ -1,4 +1,1 @@
-export * from './http';
 export * from './icons';
-export * from './models';
-export * from './try-catch';
