--- apps/web-app/src/app/shared/models/index.ts
@@ -1,1 +1,0 @@
-export * from './item.model';
