--- apps/web-app/src/app/shared/models/item.model.ts
@@ -1,6 +1,0 @@
-export interface Item {
-    id: string;
-    name: string;
-    weight: number;
-    baseValue: number;
-}
