+++ libs/shared-ui/src/lib/config/index.ts
@@ -1,0 +1,2 @@
+export * from './config.service';
+export * from './models';
