+++ libs/shared-ui/test/index.ts
@@ -1,1 +1,2 @@
 export * from './mocks';
+export * from './utils';
