+++ package.json
@@ -16,9 +16,7 @@
         "libs/*"
     ],
     "msw": {
-        "workerDirectory": [
-            "apps/web-app/.msw",
-        ]
+        "workerDirectory": ["apps/web-app/.msw"]
     },
     "dependencies": {
         "@angular/common": "~20.1.0",
