+++ tsconfig.base.json
@@ -22,6 +22,7 @@
         "noUnusedParameters": true,
         "paths": {
             "@craftsmans-ledger/shared-ui": ["libs/shared-ui/src/index.ts"],
+            "@craftsmans-ledger/shared-ui/test": ["libs/shared-ui/test/index.ts"]
         },
         "rootDir": ".",
         "skipDefaultLibCheck": true,
